# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API De Tuna Resort",
        "description": "Dokumentasi API untuk Sistem Informasi & Pemesanan De Tuna Resort",
        "contact": {
            "email": "admin@detuna.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server API De Tuna (Lokal - php artisan serve)"
        },
        {
            "url": "http://detuna-api.test",
            "description": "Server API De Tuna (Lokal - Laragon)"
        }
    ],
    "paths": {
...
        "/api/v1/rooms/{id}": {
            "get": {
                "tags": [
                    "Rooms (Public)"
                ],
                "summary": "Get Detail Kamar",
                "operationId": "20e96d2b586955bc2a4d81fdbff8789d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail kamar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomWithRelations"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kamar tidak ditemukan"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Room": {
                "title": "Room",
                "description": "Model Tipe Kamar",
                "properties": {
                    "room_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Deluxe Ocean View"
                    },
                    "description": {
                        "type": "string",
                        "example": "Kamar indah dengan pemandangan laut."
                    },
                    "price_per_night": {
                        "type": "number",
                        "format": "decimal",
                        "example": 2500000
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 2
                    },
                    "size": {
                        "type": "string",
                        "example": "45 m²"
                    },
                    "bed_type": {
                        "type": "string",
                        "example": "1 King Bed"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "maintenance"
                        ],
                        "example": "available"
                    }
                },
                "type": "object"
            },
            "RoomWithRelations": {
                "title": "Room With Relations",
                "description": "Model Kamar dengan relasi foto dan fasilitas",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Room"
                    },
                    {
                        "properties": {
                            "photos": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/RoomPhoto"
                                }
                            },
                            "facilities": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/RoomFacility"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "RoomFacility": {
                "title": "Room Facility",
                "description": "Model Fasilitas di dalam Kamar",
                "properties": {
                    "facility_id": {
                        "type": "integer",
                        "example": 201
                    },
                    "room_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "facility_name": {
                        "type": "string",
                        "example": "Wi-Fi Gratis"
                    }
                },
                "type": "object"
            },
            "RoomPhoto": {
                "title": "Room Photo",
                "description": "Model Foto Kamar",
                "properties": {
                    "photo_id": {
                        "type": "integer",
                        "example": 101
                    },
                    "room_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "photo_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://example.com/images/room1.jpg"
                    },
                    "is_primary": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token (Bearer Token)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "sanctum": {
            "type": "http",
            "description": "Masukkan Bearer Token (Sanctum)",
            "scheme": "bearer",
            "bearerFormat": "JWT"
        }
    }
}